requirements.txt
toml~=0.10.1
pyte~=0.8.0
discord.py==2.3.2
requests